requirements.txt
asyncio==3.4.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pip==23.2.1
platformdirs==3.9.1
requests==2.31.0
setuptools==68.0.0
urllib3==2.0.4
waitress==2.1.2
websockets==11.0.3
Werkzeug==2.3.6
wheel==0.41.0

Pillow~=10.0.0
roman~=4.1